Add documentation for unsafe.rkt
The unsafe.rkt module has some extra cruft that comes only from this project.
This commit makes a more serious effort to explain what's going on in unsafe.rkt
on instantiation, and what exactly is provided.

This includes some platform-specific C types, hand-maintained procedures,
and a clarification of the module's purpose in the greater scheme of things.