Remove auto-checking VkResult by default
This is the first step towards fulfilling exposure of a raw C API
with opt-in conveniences. If you want to automatically check return
codes, you have to say so with a command line.

This does bring about a breaking change in `unsafe.rkt`, but the
prior code is reproducible. The docs still warn that the interface
is unstable, so that's my excuse for doing stuff like this until
I hit 1.0.